requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-telegram-bot[all]==21.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
redis[hiredis]==5.1.0
pinecone-client==5.0.0
anthropic==0.39.0
ccxt==4.4.0
httpx==0.27.0
tavily-python==0.5.0
playwright==1.48.0
cryptography==43.0.0
python-dotenv==1.0.1
apscheduler==3.10.4
pillow==10.4.0
aiosqlite==0.20.0
aiohttp==3.10.0
